# swagger/swagger.yaml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_healthCheck
      parameters: []
      responses:
        '200':
          description: ''
  /clausulas:
    post:
      operationId: ClausulaController_post
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClausulaDto'
      responses:
        '201':
          description: ''
      tags: &ref_0
        - clausulas
    get:
      operationId: ClausulaController_getAll
      parameters:
        - name: query
          required: false
          in: query
          description: query - Filter. e.g. col1:v1,col2:v2 …
          schema:
            type: string
        - name: fields
          required: false
          in: query
          description: fields - Fields returned. e.g. col1,col2 …
          schema:
            type: string
        - name: sortby
          required: false
          in: query
          description: sortby - Sorted-by fields. e.g. col1,col2 …
          schema:
            type: string
        - name: order
          required: false
          in: query
          description: >-
            order - Order corresponding to each sortby field, if single value,
            apply to all sortby fields. e.g. desc,asc …
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: limit - Limit the size of result set. Must be an integer
          schema:
            type: string
        - name: offset
          required: false
          in: query
          description: offset - Start position of result set. Must be an integer
          schema:
            type: string
        - name: populate
          required: false
          in: query
          description: populate - show subqueries. Must be a boolean
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_0
  /clausulas/{id}:
    get:
      operationId: ClausulaController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_0
    put:
      operationId: ClausulaController_put
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClausulaDto'
      responses:
        '200':
          description: ''
      tags: *ref_0
    delete:
      operationId: ClausulaController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_0
  /paragrafos:
    post:
      operationId: ParagrafoController_post
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateParagrafoDto'
      responses:
        '201':
          description: ''
      tags: &ref_1
        - paragrafos
    get:
      operationId: ParagrafoController_getAll
      parameters:
        - name: query
          required: false
          in: query
          description: query - Filter. e.g. col1:v1,col2:v2 …
          schema:
            type: string
        - name: fields
          required: false
          in: query
          description: fields - Fields returned. e.g. col1,col2 …
          schema:
            type: string
        - name: sortby
          required: false
          in: query
          description: sortby - Sorted-by fields. e.g. col1,col2 …
          schema:
            type: string
        - name: order
          required: false
          in: query
          description: >-
            order - Order corresponding to each sortby field, if single value,
            apply to all sortby fields. e.g. desc,asc …
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: limit - Limit the size of result set. Must be an integer
          schema:
            type: string
        - name: offset
          required: false
          in: query
          description: offset - Start position of result set. Must be an integer
          schema:
            type: string
        - name: populate
          required: false
          in: query
          description: populate - show subqueries. Must be a boolean
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_1
  /paragrafos/{id}:
    get:
      operationId: ParagrafoController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_1
    put:
      operationId: ParagrafoController_put
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateParagrafoDto'
      responses:
        '200':
          description: ''
      tags: *ref_1
    delete:
      operationId: ParagrafoController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_1
  /orden-clausulas:
    post:
      operationId: OrdenClausulaController_post
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrdenClausulaDto'
      responses:
        '201':
          description: ''
      tags: &ref_2
        - orden-clausulas
    get:
      operationId: OrdenClausulaController_getAll
      parameters:
        - name: query
          required: false
          in: query
          description: query - Filter. e.g. col1:v1,col2:v2 …
          schema:
            type: string
        - name: fields
          required: false
          in: query
          description: fields - Fields returned. e.g. col1,col2 …
          schema:
            type: string
        - name: sortby
          required: false
          in: query
          description: sortby - Sorted-by fields. e.g. col1,col2 …
          schema:
            type: string
        - name: order
          required: false
          in: query
          description: >-
            order - Order corresponding to each sortby field, if single value,
            apply to all sortby fields. e.g. desc,asc …
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: limit - Limit the size of result set. Must be an integer
          schema:
            type: string
        - name: offset
          required: false
          in: query
          description: offset - Start position of result set. Must be an integer
          schema:
            type: string
        - name: populate
          required: false
          in: query
          description: populate - show subqueries. Must be a boolean
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_2
  /orden-clausulas/{id}:
    get:
      operationId: OrdenClausulaController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_2
    put:
      operationId: OrdenClausulaController_put
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrdenClausulaDto'
      responses:
        '200':
          description: ''
      tags: *ref_2
    delete:
      operationId: OrdenClausulaController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_2
  /orden-paragrafos:
    post:
      operationId: OrdenParagrafoController_post
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrdenParagrafoDto'
      responses:
        '201':
          description: ''
      tags: &ref_3
        - orden-paragrafos
    get:
      operationId: OrdenParagrafoController_getAll
      parameters:
        - name: query
          required: false
          in: query
          description: query - Filter. e.g. col1:v1,col2:v2 …
          schema:
            type: string
        - name: fields
          required: false
          in: query
          description: fields - Fields returned. e.g. col1,col2 …
          schema:
            type: string
        - name: sortby
          required: false
          in: query
          description: sortby - Sorted-by fields. e.g. col1,col2 …
          schema:
            type: string
        - name: order
          required: false
          in: query
          description: >-
            order - Order corresponding to each sortby field, if single value,
            apply to all sortby fields. e.g. desc,asc …
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: limit - Limit the size of result set. Must be an integer
          schema:
            type: string
        - name: offset
          required: false
          in: query
          description: offset - Start position of result set. Must be an integer
          schema:
            type: string
        - name: populate
          required: false
          in: query
          description: populate - show subqueries. Must be a boolean
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_3
  /orden-paragrafos/{id}:
    get:
      operationId: OrdenParagrafoController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_3
    put:
      operationId: OrdenParagrafoController_put
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrdenParagrafoDto'
      responses:
        '200':
          description: ''
      tags: *ref_3
    delete:
      operationId: OrdenParagrafoController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_3
  /plantilla-tipo-contratos:
    post:
      operationId: PlantillaTipoContratoController_post
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlantillaTipoContratoDto'
      responses:
        '201':
          description: ''
      tags: &ref_4
        - plantilla-tipo-contratos
    get:
      operationId: PlantillaTipoContratoController_getAll
      parameters:
        - name: query
          required: false
          in: query
          description: query - Filter. e.g. col1:v1,col2:v2 …
          schema:
            type: string
        - name: fields
          required: false
          in: query
          description: fields - Fields returned. e.g. col1,col2 …
          schema:
            type: string
        - name: sortby
          required: false
          in: query
          description: sortby - Sorted-by fields. e.g. col1,col2 …
          schema:
            type: string
        - name: order
          required: false
          in: query
          description: >-
            order - Order corresponding to each sortby field, if single value,
            apply to all sortby fields. e.g. desc,asc …
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: limit - Limit the size of result set. Must be an integer
          schema:
            type: string
        - name: offset
          required: false
          in: query
          description: offset - Start position of result set. Must be an integer
          schema:
            type: string
        - name: populate
          required: false
          in: query
          description: populate - show subqueries. Must be a boolean
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_4
  /plantilla-tipo-contratos/{id}:
    get:
      operationId: PlantillaTipoContratoController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_4
    put:
      operationId: PlantillaTipoContratoController_put
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlantillaTipoContratoDto'
      responses:
        '200':
          description: ''
      tags: *ref_4
    delete:
      operationId: PlantillaTipoContratoController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_4
  /plantilla-tipo-contratos/tipo-contratos/{tipo_contrato_id}:
    get:
      operationId: PlantillaTipoContratoController_getByTipoContrato
      parameters:
        - name: tipo_contrato_id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
      tags: *ref_4
    put:
      operationId: PlantillaTipoContratoController_putTipoContrato
      parameters:
        - name: tipo_contrato_id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlantillaTipoContratoDto'
      responses:
        '200':
          description: ''
      tags: *ref_4
    delete:
      operationId: PlantillaTipoContratoController_deleteTipoContrato
      parameters:
        - name: tipo_contrato_id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
      tags: *ref_4
info:
  title: Polizas API CRUD
  description: API
  version: '1.0'
  contact: {}
tags:
  - name: API
    description: ''
servers: []
components:
  schemas:
    CreateClausulaDto:
      type: object
      properties:
        nombre:
          type: string
        descripcion:
          type: string
        predeterminado:
          type: boolean
        fecha_creacion:
          format: date-time
          type: string
        fecha_modificacion:
          format: date-time
          type: string
      required:
        - nombre
        - descripcion
        - predeterminado
        - fecha_creacion
        - fecha_modificacion
    CreateParagrafoDto:
      type: object
      properties:
        nombre:
          type: string
        descripcion:
          type: string
        predeterminado:
          type: boolean
        fecha_creacion:
          format: date-time
          type: string
        fecha_modificacion:
          format: date-time
          type: string
      required:
        - nombre
        - descripcion
        - predeterminado
        - fecha_creacion
        - fecha_modificacion
    CreateOrdenClausulaDto:
      type: object
      properties:
        clausula_ids:
          description: Lista de Clausula IDs
          type: array
          items:
            type: string
        contrato_id:
          type: string
          description: Contrato ID
        fecha_creacion:
          format: date-time
          type: string
        fecha_modificacion:
          format: date-time
          type: string
      required:
        - clausula_ids
        - contrato_id
        - fecha_creacion
        - fecha_modificacion
    CreateOrdenParagrafoDto:
      type: object
      properties:
        paragrafo_ids:
          description: Lista de Paragrafo IDs
          type: array
          items:
            type: string
        contrato_id:
          type: string
          description: Contrato ID
        clausula_id:
          type: string
          description: Clausula ID
        fecha_creacion:
          format: date-time
          type: string
        fecha_modificacion:
          format: date-time
          type: string
      required:
        - paragrafo_ids
        - contrato_id
        - clausula_id
        - fecha_creacion
        - fecha_modificacion
    CreatePlantillaTipoContratoDto:
      type: object
      properties:
        version:
          type: number
        version_actual:
          type: boolean
        tipo_contrato_id:
          type: number
        orden_clausula_id:
          type: string
          description: Orden Clausula ID
        orden_paragrafo_ids:
          description: Lista de Orden Paragrafo IDs
          type: array
          items:
            type: string
        fecha_creacion:
          format: date-time
          type: string
        fecha_modificacion:
          format: date-time
          type: string
      required:
        - version
        - version_actual
        - tipo_contrato_id
        - orden_clausula_id
        - orden_paragrafo_ids
        - fecha_creacion
        - fecha_modificacion